requirements.txt
Django
django-extensions
pygraphviz
djangorestframework